parallelize blocking-moves search
This does one thread per possibility in the first level of the tree,
so it's not tuned to the computer's processing power, and if one branch
is much larger than the others, we'll be sitting and waiting for it
when we could be allowing it to bifurcate. A pool model where
execution branches anytime there are unoccupied workers in the
pool would be more powerful.